+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Updated/Enhanced Rules and Bug Fixes
    - #207: **ParameterCount** ignores methods with @Override annotation; *ignoreOverriddenMethods* defaults to true. (Thanks to Yuriy Chulovskyy)
    - #223: **UnusedVariable**: Don't count variable assignment as a reference (usage).
    - #230: **NoDef**: ClassCastException: ArgumentListExpression cannot be cast to VariableExpression.
+   - #226: **UnusedPrivateField** rule should ignore fields annotated with `groovy.lang.Delegate`.
 
 Framework and Infrastructure
    - #228: [BREAKING CHANGE] Upgrade to Groovy 2.4.